请假按钮高度处理

+++ src/components/Enclosure/enclosure.vue
@@ -4,18 +4,24 @@
     <Header title="附件管理" @clickLeft="clickLeft"></Header>
     <div class="item_body" :style="{'height': currentHeight}">
       <div v-if="fileList.length != 0">
-        <van-cell v-for="(file, index) in fileList" :key="index"  center
-                  :title-style="{'margin-left':'10px'}">
-          <template #title>
-            {{file.fileName}}
+        <van-swipe-cell v-for="(file, index) in fileList" :key="index"  :disabled="disabled">
+          <!-- 单元格-->
+          <van-cell center :title-style="{'margin-left':'10px'}">
+            <template #title>
+              {{file.fileName}}
+            </template>
+            <!-- 单元格左侧图标-->
+            <template #icon>
+              <van-icon name="card" color="#2479ed" size="19"/>
+            </template>
+          </van-cell>
+          <!-- 单元格滑动操作-->
+          <template #right>
+            <van-button square type="info" icon="eye-o" @click="lookFile(file)"/>
+            <van-button square type="primary" icon="down" @click="downFile(file)"/>
+            <van-button square type="danger" icon="delete" @click="deleteFile(file,index)"/>
           </template>
-          <template #icon>
-            <van-icon name="card" color="#2479ed" size="19" v-if="!disabled"/>
-          </template>
-          <template #right-icon>
-            <van-icon name="delete" color="#ee0a24" size="19" v-if="!disabled" @click="deleteFile(file,index)"/>
-          </template>
-        </van-cell>
+        </van-swipe-cell>
       </div>
       <!--无数据-->
       <div v-else>
@@ -33,9 +39,10 @@
 
 <script>
   import Header from '@/components/Header/Index'
-  import {getFileList,deleteFile,uploadFile} from '@/api/filemanager'
+  import {getFileList, deleteFile, uploadFile} from '@/api/filemanager'
   import {Toast} from 'vant';
-
+  import { Dialog } from 'vant';
+  import { ImagePreview } from 'vant';
   export default {
     name: 'addressList',
     data() {
@@ -48,15 +55,36 @@
       }
     },
     components: {
-      Header
+      Header,
+      [ImagePreview.Component.name]: ImagePreview.Component
     },
     mounted() {
       this.currentHeight = (document.documentElement.clientHeight - 46 - 60) + 'px'
       this.uploadButtonWidth = (document.documentElement.clientWidth - 20) + 'px'
       this.filePath = this.$route.query.filePath;
-      this.queryFileList(this.$route.query.filePath)
+      if(this.$route.query.disabled === 'true') {
+        this.disabled = true
+      }
+      this.queryFileList(this.$route.query.filePath, true)
     },
     methods: {
+      /**
+       * 下载附件
+       */
+      downFile(file){
+        window.location.href = file.downloadUrl
+      },
+      /**
+       * 查看附件
+       */
+      lookFile(file){
+        ImagePreview({
+          images: [
+            file.viewUrl
+          ],
+          closeable: true
+        });
+      },
       /**
        * 上传附件
        */
@@ -68,56 +96,44 @@
         params.append("file", file.file)
         uploadFile(params).then(res => {
           Toast.success(res.message)
-          this.queryFileList()
+          this.queryFileList(this.filePath, false)
         })
-      },
-      handleFiles(files) {
-      },
-      renderImage: function (file) {
-        let _this = this;
-        let leng = file.length;
-        for (let i = 0; i < leng; i++) {
-          let name = file[i].name
-          let reader = new FileReader();
-          reader.readAsDataURL(file[i]);
-          reader.onload = function (e) {
-            let result = e.target.result;
-            let imgObj = {
-              name: name,
-              data: result,
-            }
-            //调用上传接口上传附件
-            _this.doSave(imgObj)
-          };
-        }
-      },
-      doSave(imgObj) {
       },
       /**
        * 获取附件列表
        */
-      queryFileList(filePath) {
-        Toast.loading({
-          message: '加载中...',
-          duration: 0
-        })
+      queryFileList(filePath, isToast) {
+        if (isToast) {
+          Toast.loading({
+            message: '加载中...',
+            duration: 0
+          })
+        }
         let params = {
           filePath: filePath
         }
         getFileList(params).then(res => {
           this.fileList = res.data
-          Toast.clear()
+          if (isToast) {
+            Toast.clear()
+          }
         })
       },
       /**
        * 删除附件事件
        */
-      deleteFile(file,index) {
-        let params = []
-        params.push(file.nodePath)
-        deleteFile(params).then(res => {
-          this.fileList.splice(index, 1)
-          Toast.success(res.message)
+      deleteFile(file, index) {
+        Dialog.confirm({
+          title: '删除附件',
+          message: '确认要删除该附件吗?',
+        }).then(() => {
+          let params = []
+          params.push(file.nodePath)
+          deleteFile(params).then(res => {
+            this.fileList.splice(index, 1)
+            Toast.success(res.message)
+          })
+        }).catch(() => {
         })
       },
       clickLeft() {
